Moved internal device code out of hw/inc and improved naming consistency
