+++ tt_metal/programming_examples/CMakeLists.txt
@@ -16,7 +16,6 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/eltwise_sfpu)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/hello_world_compute_kernel)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/hello_world_datamovement_kernel)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/hello_world_datatypes_kernel)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/lazy_expressions)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/loopback)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/matmul)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/pad_multi_core)
