--- tt_metal/programming_examples/lazy_expressions/kernels/dataflow/write_tiles.cpp
@@ -1,16 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#include "views/accessor_view.h"
-
-void kernel_main() {
-    const auto n_tiles = get_arg_val<uint32_t>(0);
-    const auto start_id = get_arg_val<uint32_t>(1);
-
-    constexpr auto num_tiles_per_cycle = get_compile_time_arg_val(0);
-
-    // 1 circular buffer with bank base address at compile args offset 1 and runtime args offset 2
-    const auto view = views::AccessorView<1, 1, num_tiles_per_cycle>(2);
-    view.write_tiles(n_tiles, start_id);
-}
