+++ ttnn/cpp/ttnn/operations/eltwise/materialize/CMakeLists.txt
@@ -1,10 +1,10 @@
-add_library(ttnn_op_eltwise_fused ${LIB_TYPE})
-add_library(TTNN::Ops::Eltwise::Fused ALIAS ttnn_op_eltwise_fused)
+add_library(ttnn_op_eltwise_materialize ${LIB_TYPE})
+add_library(TTNN::Ops::Eltwise::Materialize ALIAS ttnn_op_eltwise_materialize)
 
-target_precompile_headers(ttnn_op_eltwise_fused REUSE_FROM TT::CommonPCH)
-TT_ENABLE_UNITY_BUILD(ttnn_op_eltwise_fused)
+target_precompile_headers(ttnn_op_eltwise_materialize REUSE_FROM TT::CommonPCH)
+TT_ENABLE_UNITY_BUILD(ttnn_op_eltwise_materialize)
 set_target_properties(
-    ttnn_op_eltwise_fused
+    ttnn_op_eltwise_materialize
     PROPERTIES
         INTERFACE_HEADER_TO_VERIFY
             api
@@ -13,7 +13,7 @@ set_target_properties(
 # Globbing non-build files is acceptable for now because devs don't generate packages.
 file(GLOB_RECURSE kernels device/kernels/*)
 target_sources(
-    ttnn_op_eltwise_fused
+    ttnn_op_eltwise_materialize
     PUBLIC
         FILE_SET api
         TYPE HEADERS
@@ -28,9 +28,9 @@ target_sources(
         device/materialize_program_factory.cpp
 )
 
-target_include_directories(ttnn_op_eltwise_fused PRIVATE ${FixmeOpIncDirs})
+target_include_directories(ttnn_op_eltwise_materialize PRIVATE ${FixmeOpIncDirs})
 target_link_libraries(
-    ttnn_op_eltwise_fused
+    ttnn_op_eltwise_materialize
     PRIVATE
         TT::Metalium
         TTNN::Core
@@ -38,10 +38,10 @@ target_link_libraries(
 
 install(
     TARGETS
-        ttnn_op_eltwise_fused
+        ttnn_op_eltwise_materialize
     FILE_SET
     kernels
-        DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/tt-metalium/ttnn/cpp/ttnn/operations/eltwise/fused
+        DESTINATION ${CMAKE_INSTALL_LIBEXECDIR}/tt-metalium/ttnn/cpp/ttnn/operations/eltwise/materialize
         COMPONENT ttnn-runtime
 )
-install(TARGETS ttnn_op_eltwise_fused LIBRARY COMPONENT tar)
+install(TARGETS ttnn_op_eltwise_materialize LIBRARY COMPONENT tar)
