+++ ttnn/cpp/ttnn/operations/eltwise/materialize/materialize_pybind.hpp
@@ -6,10 +6,10 @@
 
 #include "ttnn-pybind/pybind_fwd.hpp"
 
-namespace ttnn::operations::fused {
+namespace ttnn::operations::materialize {
 
 namespace py = pybind11;
 
 void py_module(py::module& module);
 
-}  // namespace ttnn::operations::fused
+}  // namespace ttnn::operations::materialize
